Progress made not fully complete
I have a profile section and a new nav bar, moved the logout button too the top right corner and the nofitcation bell to the top left corner. Having an issue I might need help with, when a user logouts there an infinite redirect issue happening inside of the nav bar. really weird. Just have to update the dashboard icons with bootstrap ones and fix that redirect issue and then that's everything.