Final Changes
Everything is complete, profile section, updated bootstrap nav bar and icons. Only problem is that when a user goes to the login screen for some reason the page loads inside of the nav bar, causing to instances of the page to be open at once. You can refresh the page and it solves the issue.